Speed up bitmask operations
