+++ encord/objects/bitmask.py
@@ -15,6 +15,7 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional, Protocol, Sequence, Tuple, Union, runtime_checkable
 
 from encord.exceptions import EncordException
+from encord.objects.bitmask_operations import deserialise_bitmask, serialise_bitmask, transpose_bytearray
 from encord.orm.base_dto import BaseDTO
 
 
@@ -29,123 +30,6 @@ class ArrayProtocol(Protocol):
     def __array_interface__(self) -> Dict[str, Any]: ...
 
     def tobytes(self) -> bytes: ...
-
-
-def _string_to_rle(s: str) -> List[int]:
-    """
-    COCO-compatible string to RLE-encoded mask de-serialisation
-    """
-    cnts: List[int] = []
-    p = 0
-
-    while p < len(s):
-        x = 0
-        k = 0
-        more = 1
-
-        while more and p < len(s):
-            c = ord(s[p]) - 48
-            x |= (c & 0x1F) << (5 * k)
-            more = c & 0x20
-            p += 1
-            k += 1
-
-            if not more and (c & 0x10):
-                x |= -1 << (5 * k)
-
-        if len(cnts) > 2:
-            x += cnts[-2]
-
-        cnts.append(x)
-
-    return cnts
-
-
-def _rle_to_string(rle: Sequence[int]) -> str:
-    """
-    COCO-compatible RLE-encoded mask to string serialisation
-    """
-    rle_string = ""
-    for i, x in enumerate(rle):
-        if i > 2:
-            x -= rle[i - 2]
-
-        more = 1
-        while more:
-            c = x & 0x1F
-            x >>= 5
-
-            if c & 0x10:
-                more = x != -1
-            else:
-                more = x != 0
-
-            if more:
-                c |= 0x20
-
-            c += 48
-            rle_string += chr(c)
-
-    return rle_string
-
-
-def _rle_to_mask(rle: List[int], size: int) -> bytes:
-    """
-    COCO-compatible RLE to bitmask
-    """
-    res = bytearray(size)
-    offset = 0
-
-    for i, c in enumerate(rle):
-        v = i % 2
-        while c > 0:
-            res[offset] = v
-            offset += 1
-            c -= 1
-
-    return bytes(res)
-
-
-def _mask_to_rle(mask: bytes) -> List[int]:
-    """
-    COCO-compatible raw bitmask to COCO-compatible RLE
-    """
-    rle_counts = []
-    c = 0
-    p = 0
-    for mask_value in mask:
-        if mask_value != p:
-            rle_counts.append(c)
-            c = 0
-            p = mask_value
-        c += 1
-
-    rle_counts.append(c)
-    return rle_counts
-
-
-def transpose_bytearray(byte_data: bytes, shape: Tuple[int, int]) -> bytearray:
-    """
-    Transpose a 2D array represented by bytes.
-    """
-    np_found = True
-    try:
-        import numpy as np
-    except ImportError:
-        np_found = False
-    if not np_found:
-        rows, cols = shape
-        # Create a new bytearray to hold the transposed data
-        transposed_byte_data = bytearray(len(byte_data))
-        # Transpose the 2D array
-        for row in range(rows):
-            for col in range(cols):
-                transposed_byte_data[col * rows + row] = byte_data[row * cols + col]
-    else:
-        np_byte_data = np.frombuffer(byte_data, dtype=np.int8).reshape(shape)
-        transposed_byte_data = bytearray(np_byte_data.T.tobytes())
-
-    return transposed_byte_data
 
 
 class BitmaskCoordinates:
@@ -212,8 +96,7 @@ class BitmaskCoordinates:
 
         raw_data = data if isinstance(data, bytes) else source.tobytes()
 
-        rle = _mask_to_rle(raw_data)
-        rle_string = _rle_to_string(rle)
+        rle_string = serialise_bitmask(raw_data)
 
         return BitmaskCoordinates.EncodedBitmask(top=0, left=0, height=shape[0], width=shape[1], rle_string=rle_string)
 
@@ -241,8 +124,9 @@ class BitmaskCoordinates:
 
     @property
     def __array_interface__(self):
-        rle = _string_to_rle(self._encoded_bitmask.rle_string)
-        data = _rle_to_mask(rle, self._encoded_bitmask.height * self._encoded_bitmask.width)
+        data = deserialise_bitmask(
+            self._encoded_bitmask.rle_string, self._encoded_bitmask.height * self._encoded_bitmask.width
+        )
         return {
             "version": 3,
             "data": data,
